fix: address remaining review items for merge readiness
- models.go: change FileSize from int to int64 across all three models
- group_document.go: add safe type assertion for userID.(uint)
- convert.go: document Linux-only syscall constraint
- main.go: document WriteTimeout 120s tradeoff
- Update all FileSize assignment sites to int64 cast

+++ internal/handlers/animal_image.go
@@ -219,7 +219,7 @@ func UploadAnimalImageToGallery(db *gorm.DB, storageProvider storage.Provider) g
 			Caption:         caption,
 			Width:           finalBounds.Dx(),
 			Height:          finalBounds.Dy(),
-			FileSize:        len(imageData),
+			FileSize:        int64(len(imageData)),
 			StorageProvider: storageProviderName,
 			BlobIdentifier:  blobIdentifier,
 			BlobExtension:   blobExt,
